Refactor Data Handling and Plotting Functions.
I stored the parameters of events in a pandas DataFrame named df_events, as an inheritance of the main class.
I separated the creation of df_stations and df_phases into two distinct functions.
I modified the plot_station_participation_per_event() function to retrieve the num_stations column and used methods of the Pandas Series to plot the barplot.